+++ Changelog.md
@@ -1,5 +1,14 @@
 # Westwind.HtmlPackager
 
+### 0.1.4
+*November 1st, 2018*
+
+* **Add display switch to command line processor**  
+You can now use the `-d` switch to preview the generated HTML or zip file.
+
+* **Add Zip File Packaging**  
+You can now package the generated HTML file into a self-contained zip file using the `-z` switch.
+
 ### 0.1.3
 *September 8th, 2018*